Fix Mac Slicon Problems
* Fix mp3, wav selection.
* Fix float64 error on apple slicons.

+++ vc_infer_pipeline.py
@@ -330,7 +330,7 @@ class VC(object):
             pitch = pitch[:p_len]
             pitchf = pitchf[:p_len]
             pitch = torch.tensor(pitch, device=self.device).unsqueeze(0).long()
-            pitchf = torch.tensor(pitchf, device=self.device).unsqueeze(0).float()
+            pitchf = torch.tensor(pitchf, device=self.device, dtype=torch.float32).unsqueeze(0).float()
         t2 = ttime()
         times[1] += t2 - t1
         for t in opt_ts:
